task_manager/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-debug-toolbar
